feat: 예약 취소시 회원 로그인 상태 및 권한 확인 기능 추가

+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -28,7 +28,7 @@ public class ReservationController {
         if (token == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-        memberService.findByToken(token);
+        Member member = memberService.findByToken(token);
         Long id = reservationService.create(reservationRequest);
         return ResponseEntity.created(URI.create("/reservations/" + id)).build();
     }
@@ -40,7 +40,18 @@ public class ReservationController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable Long id) {
+    public ResponseEntity deleteReservation(@AuthenticationPrincipal String token, @PathVariable Long id) {
+        if (token == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        // 권한 체크(본인의 예약건인지 확인)
+        Member member = memberService.findByToken(token);
+        Reservation reservation = reservationService.findById(id);
+        if (!reservation.getName().equals(member.getName())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         reservationService.deleteById(id);
 
         return ResponseEntity.noContent().build();
